requirements.txt
datasets>=2.20.0
pandas>=2.0.0
pydantic>=2.7.0
tqdm>=4.66.0
openai>=1.40.0
python-dotenv>=1.0.0
loguru>=0.7.0
